Large changes: **WARNING** your code will probably break if you use the graphical interface.
If you need the old version, use the CVS tag END_VERSION_0_5

The version has been bumped to 0.6, so it shouldn't overwrite the old 0.5
library, so compiled programs shouldn't break (hpefully).

Changes:

1 The GUI_{motif,fltk,fltk2} interface has been reworked.

  You should now include GUI_Widgets.h. This has a class called GUI_Widgets.h
  with two methods: run_thread() and process_in_crnt_thread(). This does the
  same as before, except that your code will now always compile, regardless of
  the type of widget set available.

  Also, all GUI calls are via GUI.*, not GUI_{motif,fltk,fltk2}.*, so you won't
  have to rewrite your .cfg files if you have a different library set available.

2 GUI_Motif now supports process_in_crnt_thread(), so the mysterious segfaults
  have probably vanished. YMMV

3 GUI_Fltk has been converted to GVars3. fltk2 and motif still use the backwards
  compatibility interface

4 The configure script now selects only one of the available widget sets to
  compile, choosing (in order) fltk2, fltk, motif, none. There are various
  enable/disable options.


With any luck, programs should "just work" rather better than they used to, and
the interface is simpler.

It currently uses some implementation defined behaviour in g++, that static
constructres are run in the reverse order of object files on the commandline.