A few basic language constructs. This is an excellent illustration as to why language design should be left to computer scientists.
Examples:

function foo
. echo hello
. echo world
. a=4
endfunction

foo

if_equal a 4
. echo yes, a is 4
else
. echo dang
. if_equal a 3
. . echo OK, a is 3
. endif
endif

Bear in mind that . is simply a perfectly normal registered command.

Essentially, "function" is a command that says "rember stuff"
"." shoves a line in to a list of lines from the last rememberes point
"endfunction" then saves the remembered stuff permenantly and registeres
a command to ParseLine it one line at a time.

if is similar.

Now imagine what happens if you interleave scoping bits. Or how the quoting
works (how many times is a particular line evaluated...?)

It's thread safe, although some bits of GUI it calls aren't.